removed a unneeded cast to int of a float value.
The intensity threshold is a float in the parameters and is expected to
be a float (32f) by the ipp function that uses it, but is cast through
int in signature of FindLocalMax_2D.

+++ iden/cxx/iden.cpp
@@ -196,10 +196,10 @@ Wrapper_i_plu * Iden::fill_wrapper(unsigned int frames,unsigned int start)
     status = FindLocalMax_2D(image_bpass,
 			     image_bpass_thresh,
 			     image_localmax,
-			     (int)params_.get_pctle_threshold(),
+			     params_.get_pctle_threshold(),
 			     params_.get_dilation_radius());
     
-
+    
     
     RecenterImage(image_bpass_thresh);
     // image_in.display_image();
